% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACAT}
\alias{ACAT}
\title{ACAT function to combine SNP pvalues}
\usage{
ACAT(Pvals, weights = NULL, is.check = T, is.small.thre = 1e-15)
}
\arguments{
\item{Pvals}{Single SNP pvalues to be combined}

\item{weights}{ACAT weights}

\item{is.check}{Check pvalues}

\item{is.small.thre}{Threshold for pvalue to be adjusted}
}
\value{
ACAT combined pvalue
}
\description{
ACAT function to combine SNP pvalues
}
