% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACATQ_fun}
\alias{ACATQ_fun}
\title{SS_ACAT SNP score and test statistics function}
\usage{
ACATQ_fun(G, cvalue, mac.thresh, is.very.rare, wBurden)
}
\arguments{
\item{G}{Genotype data}

\item{cvalue}{Individual c values}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{is.very.rare}{If the SNPs are very rare. TRUE for rare.}

\item{wBurden}{Vector of SNP burden weights.}
}
\value{
SS_ACAT scores and test statistics
}
\description{
SS_ACAT SNP score and test statistics function
}
