% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACATscore_fun}
\alias{ACATscore_fun}
\title{SS_ACAT SNP score function}
\usage{
ACATscore_fun(Gw, G, cvalue, mac.thresh, is.very.rare)
}
\arguments{
\item{Gw}{Burden weighted genotype data}

\item{G}{Genotype data}

\item{cvalue}{Individual c values}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{is.very.rare}{If the SNPs are very rare. TRUE for rare.}
}
\value{
SS_ACAT scores
}
\description{
SS_ACAT SNP score function
}
