% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACATsingleQ_fun}
\alias{ACATsingleQ_fun}
\title{SS_ACAT single SNP score and test statistics function}
\usage{
ACATsingleQ_fun(G, cvalue)
}
\arguments{
\item{G}{Genotype data}

\item{cvalue}{Individual c values}
}
\value{
SS_ACAT single SNP scores and test statistics
}
\description{
SS_ACAT single SNP score and test statistics function
}
