% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACATsinglescore_fun}
\alias{ACATsinglescore_fun}
\title{SS_ACAT single SNP score function}
\usage{
ACATsinglescore_fun(G, cvalue)
}
\arguments{
\item{G}{Genotype data}

\item{cvalue}{Individual c values}
}
\value{
SS_ACAT single SNP score
}
\description{
SS_ACAT single SNP score function
}
