% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACATSsingleVar_fun}
\alias{ACATSsingleVar_fun}
\title{SS_ACAT single SNP score function variances}
\usage{
ACATSsingleVar_fun(G, para_cvalue)
}
\arguments{
\item{G}{Genotype data}

\item{para_cvalue}{Parametric boostrapped individual c values}
}
\value{
SS_ACAT single SNP variances
}
\description{
SS_ACAT single SNP score function variances
}
