% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACATSVar_fun}
\alias{ACATSVar_fun}
\title{SS_ACAT SNP score variance function}
\usage{
ACATSVar_fun(G, para_cvalue, is.very.rare, mac.thresh, wBurden)
}
\arguments{
\item{G}{Genotype data}

\item{para_cvalue}{Parametric boostrapped individual c values}

\item{is.very.rare}{If the SNPs are very rare.}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{wBurden}{Vector of SNP burden weights.}
}
\value{
SS_ACAT variance
}
\description{
SS_ACAT SNP score variance function
}
