% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ACATW_func}
\alias{ACATW_func}
\title{Weights function for ACAT}
\usage{
ACATW_func(G, is.very.rare, wBurden, weights.beta)
}
\arguments{
\item{G}{Genotype matrix.}

\item{is.very.rare}{If the SNPs are very rare. TRUE for rare.}

\item{wBurden}{Vector of SNP burden weights. When it is NULL, the beta weight with the “weights.beta” parameter is used.}

\item{weights.beta}{Weights beta parameters.}
}
\value{
ACAT weights
}
\description{
Weights function for ACAT
}
