% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Burden}
\alias{Burden}
\title{Chi-square for SS_Burden pvalue calculation}
\usage{
Burden(S, sigma)
}
\arguments{
\item{S}{S score of burden, Q=S'S.}

\item{sigma}{Variance of S.}
}
\value{
pvalue.
}
\description{
Chi-square for SS_Burden pvalue calculation
}
