% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{BurdenQ_fun}
\alias{BurdenQ_fun}
\title{SS_burden score and test statistics function}
\usage{
BurdenQ_fun(G, cvalue, wBurden)
}
\arguments{
\item{G}{Genotype data}

\item{cvalue}{Individual c values}

\item{wBurden}{Vector of SNP burden weights.}
}
\value{
SS_burden score and test statistics
}
\description{
SS_burden score and test statistics function
}
