% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Burdenscore_fun}
\alias{Burdenscore_fun}
\title{SS_burden score function}
\usage{
Burdenscore_fun(Gw, cvalue)
}
\arguments{
\item{Gw}{Burden weighted genotype data}

\item{cvalue}{Individual c values}
}
\value{
SS_burden score
}
\description{
SS_burden score function
}
