% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{BurdenSVar_fun}
\alias{BurdenSVar_fun}
\title{SS_Burden score variance function}
\usage{
BurdenSVar_fun(G, para_cvalue, wBurden)
}
\arguments{
\item{G}{Genotype data}

\item{para_cvalue}{Parametric boostrapped individual c values}

\item{wBurden}{Vector of SNP burden weights.}
}
\value{
SS_Burden variance matrix
}
\description{
SS_Burden score variance function
}
