% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{c_func}
\alias{c_func}
\title{c value calculation to be used in score function}
\usage{
c_func(Y, X, S, Z, id.t, theta)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{S}{Continuous surrogate}

\item{Z}{Covaraites with intercept column design matrix}

\item{id.t}{Row id of labeled data.}

\item{theta}{Distribution parameters}
}
\value{
c value to be used in score function
}
\description{
c value calculation to be used in score function
}
