% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeled_test.R
\name{labeled_test}
\alias{labeled_test}
\title{Standard test function using the labeled data only}
\usage{
labeled_test(
  Y,
  X,
  G,
  id.t,
  wBurden = NULL,
  wSKAT = NULL,
  wACAT = NULL,
  mac.thresh = 10,
  testtype = "all"
)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{G}{Genotype data}

\item{id.t}{Row id of labeled data in the dataset.}

\item{wBurden}{Vector of SNP burden weights.}

\item{wSKAT}{Vector of SNP SKAT weights.}

\item{wACAT}{Vector of SNP ACAT weights.}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{testtype}{Type of test, default is "all", can also be "SKAT", "Burden", "ACAT".}
}
\value{
Parameter estimates, SS SKAT score and SS SKAT pvalue.
}
\description{
Standard test function using the labeled data only
}
\examples{
analysis_data <- example_data$my_dat;
labeled_data_id <- example_data$id.t;
labeled_SKAT_result <- labeled_SKAT(dat = analysis_data, id.t = labeled_data_id);
}
