% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Liu_adj}
\alias{Liu_adj}
\title{Liu's non-central chi-square approximation with tail adjust}
\usage{
Liu_adj(A, X_Sigma, X_mu, X)
}
\arguments{
\item{A}{The weight matrix in test statistics Q=S'AS.}

\item{X_Sigma}{Covariance matrix of the score function S.}

\item{X_mu}{Mean of the score function.}

\item{X}{Score vector S.}
}
\value{
pvalue, non-central chi-square df and nc.
}
\description{
Liu's non-central chi-square approximation with tail adjust
}
