% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_S_SKAT.R
\name{naive_SKAT}
\alias{naive_SKAT}
\title{SKAT directly with S for both labeled and unlabeled data}
\usage{
naive_SKAT(dat)
}
\arguments{
\item{dat}{Basis matrix of data set to be analyzed.}
}
\value{
Parameter estimates, SS SKAT score and SS SKAT pvalue.
analysis_data <- example_data$my_dat;
labeled_data_id <- example_data$id.t;
naive_SKAT_results <- naive_SKAT(dat = analysis_data);
}
\description{
SKAT directly with S for both labeled and unlabeled data
}
