% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_S_test.R
\name{naive_test}
\alias{naive_test}
\title{test directly with S for both labeled and unlabeled data}
\usage{
naive_test(
  X,
  G,
  S,
  id.t,
  wBurden = NULL,
  wSKAT = NULL,
  wACAT = NULL,
  mac.thresh = 10,
  testtype = "all"
)
}
\arguments{
\item{X}{Covariates}

\item{G}{Genotype data}

\item{S}{Continuous surrogate}

\item{id.t}{Row id of labeled data in the dataset.}

\item{wBurden}{Vector of SNP burden weights.}

\item{wSKAT}{Vector of SNP SKAT weights.}

\item{wACAT}{Vector of SNP ACAT weights.}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{testtype}{Type of test, default is "all", can also be "SKAT", "Burden", "ACAT".}

\item{Y}{Binary response variable}
}
\value{
Naive pvalues.
}
\description{
test directly with S for both labeled and unlabeled data
}
