% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{NULL_log_like}
\alias{NULL_log_like}
\title{Full log-likelihood function under the null}
\usage{
NULL_log_like(Y, X, S, Z, id.t, theta, distri)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{S}{Continuous surrogate}

\item{Z}{Covaraites with intercept column design matrix}

\item{id.t}{Row id of labeled data.}

\item{theta}{Distribution parameters.}

\item{distri}{Distribution of S|Y, either "normal" or "beta"}
}
\value{
log-likelihood function under the null
}
\description{
Full log-likelihood function under the null
}
