% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{NULL_nlog_like}
\alias{NULL_nlog_like}
\title{Negative log-likelihood function under the null}
\usage{
NULL_nlog_like(Y, X, S, Z, id.t, theta)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{S}{Continuous surrogate}

\item{Z}{Covaraites with intercept column design matrix}

\item{id.t}{Row id of labeled data.}

\item{theta}{Distribution parameters.}
}
\value{
Negative log-likelihood function under the null
}
\description{
Negative log-likelihood function under the null
}
