% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_func.R
\name{para_func}
\alias{para_func}
\title{Parametric boostrap function for individual c values}
\usage{
para_func(nn, theta, Y, X, S, Z, id.t, distri)
}
\arguments{
\item{nn}{Number of samples}

\item{theta}{Distribution parameters}

\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{S}{Continuous surrogate}

\item{Z}{Covaraites with intercept column design matrix}

\item{id.t}{Row id of labeled data.}

\item{distri}{Distribution of S|Y, either "normal" or "beta"}
}
\value{
Parametric boostrapped individual c values
}
\description{
Parametric boostrap function for individual c values
}
