% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plof_SS.R
\name{plof_SS}
\alias{plof_SS}
\title{plof test function}
\usage{
plof_SS(
  chr,
  gene_name,
  genofile,
  obj_SS,
  genes,
  QC_label = "annotation/info/QC_label",
  variant_type = c("SNV", "Indel", "variant"),
  geno_missing_imputation = c("mean", "minor"),
  Annotation_dir = "annotation/info/FunctionalAnnotation",
  Annotation_name_catalog,
  Use_annotation_weights = c(TRUE, FALSE),
  Annotation_name = NULL,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 1,
  mac.thresh = 10,
  boot = T,
  para_results,
  distri
)
}
\arguments{
\item{chr}{Chromosome number}

\item{gene_name}{Gene name}

\item{genofile}{Genotype file}

\item{obj_SS}{Null object}

\item{genes}{Gene list}

\item{QC_label}{Quality control label}

\item{variant_type}{Variant type}

\item{geno_missing_imputation}{Genotype missing imputation}

\item{Annotation_dir}{Annotation directory}

\item{Annotation_name_catalog}{Annotation name catalog}

\item{Use_annotation_weights}{Use annotation weights}

\item{Annotation_name}{Annotation name}

\item{rare_maf_cutoff}{Rare MAF cutoff}

\item{rv_num_cutoff}{RV number cutoff}

\item{mac.thresh}{MAC threshold}

\item{boot}{Parametric Bootstrapping}

\item{para_results}{Saved parametric Bootstrapping results}

\item{distri}{Distribution of S|Y, either "normal" or "beta"}
}
\value{
All comparison results
}
\description{
plof test function
}
