% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rare_func}
\alias{rare_func}
\title{Rareness filtering function for ACAT}
\usage{
rare_func(G, mac.thresh)
}
\arguments{
\item{G}{Genotype data}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}
}
\value{
If the SNPs are very rare. TRUE for rare.
}
\description{
Rareness filtering function for ACAT
}
