% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SKATBurdenW_func}
\alias{SKATBurdenW_func}
\title{Weights function for Burden or SKAT}
\usage{
SKATBurdenW_func(G, weights.beta)
}
\arguments{
\item{G}{Genotype matrix.}

\item{weights.beta}{Weights beta parameters.}
}
\value{
Burden or SKAT weights
}
\description{
Weights function for Burden or SKAT
}
