% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SKATQ_fun}
\alias{SKATQ_fun}
\title{SS_SKAT score and test statistics function}
\usage{
SKATQ_fun(G, cvalue, wSKAT)
}
\arguments{
\item{G}{Genotype data}

\item{cvalue}{Individual c values}

\item{wSKAT}{Vector of SNP SKAT weights.}
}
\value{
SS_SKAT score and test statistics
}
\description{
SS_SKAT score and test statistics function
}
