% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SKATscore_fun}
\alias{SKATscore_fun}
\title{SS_SKAT score function}
\usage{
SKATscore_fun(GW, cvalue)
}
\arguments{
\item{GW}{SKAT weighted genotype data}

\item{cvalue}{Individual c values}
}
\value{
SS_SKAT score
}
\description{
SS_SKAT score function
}
