% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SKATSVar_fun}
\alias{SKATSVar_fun}
\title{SS_SKAT score variance function}
\usage{
SKATSVar_fun(G, para_cvalue, wSKAT)
}
\arguments{
\item{G}{Genotype data}

\item{para_cvalue}{Parametric boostrapped individual c values}

\item{wSKAT}{Vector of SNP SKAT weights.}
}
\value{
SS_SKAT variance matrix
}
\description{
SS_SKAT score variance function
}
