% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_SKAT.R
\name{SS_SKAT}
\alias{SS_SKAT}
\title{Semi-supervised SKAT function}
\usage{
SS_SKAT(
  dat,
  id.t,
  weights = NULL,
  full_NR_evaluation = TRUE,
  nit = NULL,
  NULL_nlog_like,
  prev_est,
  nboot
)
}
\arguments{
\item{dat}{Basis matrix of data set to be analyzed.}

\item{id.t}{Row id of labeled data in the dat dataset.}

\item{weights}{Weights of the data, default treat data equally.}

\item{full_NR_evaluation}{Full optimization iteration, default to be TRUE.}

\item{nit}{Number of iteration for optimization if full_NR_evaluation is FALSE.}

\item{NULL_nlog_like}{Function to be optimized, predefined, no need to change}

\item{prev_est}{Estimate of prevalence.}

\item{nboot}{Number of bootstrap of pvalue calculate.}
}
\value{
Parameter estimates, SS SKAT score and SS SKAT pvalue.
}
\description{
Semi-supervised SKAT function
}
\examples{
analysis_data <- example_data$my_dat;
labeled_data_id <- example_data$id.t;
SS_SKAT_results <- SS_SKAT(dat = analysis_data, id.t = labeled_data_id, weights = NULL, full_NR_evaluation = TRUE, NULL_nlog_like = NULL_nlog_like, prev_est = mean(analysis_data$Y[labeled_data_id]), nboot = 50);
}
