% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_test.R
\name{SS_test}
\alias{SS_test}
\title{Semi-supervised association test function}
\usage{
SS_test(
  Y,
  X,
  G,
  S,
  id.t,
  para_results,
  theta_est,
  wBurden = NULL,
  wSKAT = NULL,
  wACAT = NULL,
  weights.beta = NULL,
  mac.thresh = 10,
  full_NR_evaluation = TRUE,
  nit = NULL,
  NULL_nlog_like,
  testtype = "all",
  boot = T
)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{G}{Genotype data}

\item{S}{Continuous surrogate}

\item{id.t}{Row id of labeled data in the dataset.}

\item{para_results}{Saved parametric boostrapped individual c value matrix.}

\item{theta_est}{Estimated parameters}

\item{wBurden}{Vector of SNP burden weights.}

\item{wSKAT}{Vector of SNP SKAT weights.}

\item{wACAT}{Vector of SNP ACAT weights.}

\item{weights.beta}{Weights of the SNPs, default NULL treat the SNPs equally, else take parameters of beta distribution.}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{full_NR_evaluation}{Full optimization iteration, default to be TRUE.}

\item{nit}{Number of iteration for optimization if full_NR_evaluation is FALSE.}

\item{NULL_nlog_like}{Function to be optimized, predefined, no need to change}

\item{testtype}{Type of test, default is "all", can also be "SKAT", "Burden", "ACAT".}

\item{boot}{Whether perform parametric bootstrap, default is TRUE.}
}
\value{
Total sample size, parameter estimates, SS scores, SS score variance, SS pvalues, weights used in each test and weights.beta.
}
\description{
Semi-supervised association test function
}
