% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl_theta.R
\name{ssl_theta}
\alias{ssl_theta}
\title{Final parameter estimates under the null}
\usage{
ssl_theta(Y, X, S, Z, id.t, weights = NULL, full_eval = TRUE, nit)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{S}{Continuous surrogate}

\item{Z}{Covaraites with intercept column design matrix}

\item{id.t}{Row id of labeled data.}

\item{weights}{Weights of the individual in the sample, default treat data equally.}

\item{full_eval}{Full optimization iteration, default to be TRUE.}

\item{nit}{Number of iteration for optimization if full_NR_evaluation is FALSE.}
}
\value{
Final estimates of the parameters
}
\description{
Final parameter estimates under the null
}
