% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAAR_SS.R
\name{STAAR_SS}
\alias{STAAR_SS}
\title{STAAR_SS function}
\usage{
STAAR_SS(
  genotype,
  phen,
  surro,
  covar,
  id.t,
  annotation_phred = NULL,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff,
  mac.thresh,
  boot = T,
  para_results
)
}
\arguments{
\item{genotype}{Genotype matrix}

\item{phen}{Phenotype}

\item{surro}{Surrogate}

\item{covar}{Covariate}

\item{id.t}{Labeled id}

\item{annotation_phred}{Annotation phred}

\item{rare_maf_cutoff}{Rare MAF cutoff}

\item{rv_num_cutoff}{RV number cutoff}

\item{mac.thresh}{MAC threshold}

\item{boot}{Parametric Bootstrapping}

\item{para_results}{Saved parametric Bootstrapping results}
}
\value{
STAAR SS, naive, thesholded, labeled, oracle results
}
\description{
STAAR_SS function
}
