% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_SKAT.R
\name{threshold_SKAT}
\alias{threshold_SKAT}
\title{threshold SKAT function using the labeled data and thresholded unlabeled data}
\usage{
threshold_SKAT(dat, id.t, thre_value)
}
\arguments{
\item{dat}{Basis matrix of data set to be analyzed.}

\item{id.t}{Row id of labeled data in the dat dataset.}

\item{thre_value}{The Threshold value for the surrogate.}
}
\value{
Parameter estimates, SS SKAT score and SS SKAT pvalue.
}
\description{
threshold SKAT function using the labeled data and thresholded unlabeled data
}
\examples{
analysis_data <- example_data$my_dat;
labeled_data_id <- example_data$id.t;
Y_threshold <- get_Y_threshold(dat = analysis_data, prev_value = mean(analysis_data$Y[labeled_data_id]));
threshold_SKAT_results <- threshold_SKAT(dat = analysis_data, id.t = labeled_data_id, thre_value = Y_threshold);
}
