% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_test.R
\name{threshold_test}
\alias{threshold_test}
\title{threshold function using the labeled data and thresholded unlabeled data}
\usage{
threshold_test(
  Y,
  X,
  G,
  S,
  id.t,
  thre_value,
  wBurden = NULL,
  wSKAT = NULL,
  wACAT = NULL,
  mac.thresh = 10,
  testtype = "all"
)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{G}{Genotype data}

\item{S}{Continuous surrogate}

\item{id.t}{Row id of labeled data in the dataset.}

\item{thre_value}{The Threshold value for the surrogate.}

\item{wBurden}{Vector of SNP burden weights.}

\item{wSKAT}{Vector of SNP SKAT weights.}

\item{wACAT}{Vector of SNP ACAT weights.}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{testtype}{Type of test, default is "all", can also be "SKAT", "Burden", "ACAT".}
}
\value{
Thresholded pvalues.
}
\description{
threshold function using the labeled data and thresholded unlabeled data
}
