% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Var_boot}
\alias{Var_boot}
\title{Single run of parametric bootstrap variance function for ACAT, Burden and SKAT}
\usage{
Var_boot(G, para_cvalue, is.very.rare, mac.thresh, wBurden, wSKAT, type)
}
\arguments{
\item{G}{Genotype data}

\item{para_cvalue}{Parametric boostrapped individual c values}

\item{is.very.rare}{If the SNPs are very rare.}

\item{mac.thresh}{A threshold of the minor allele count (MAC). The Burden test will be used to aggregate the SNPs with MAC less than this threshold.}

\item{wBurden}{Vector of SNP burden weights.}

\item{wSKAT}{Vector of SNP SKAT weights.}

\item{type}{Type of SS tests}
}
\value{
Bootstrapped SS score variances
}
\description{
Single run of parametric bootstrap variance function for ACAT, Burden and SKAT
}
