% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{y.p_NULL}
\alias{y.p_NULL}
\title{Expit function under the null model}
\usage{
y.p_NULL(a, b, x)
}
\arguments{
\item{a}{intercept}

\item{b}{slope}

\item{x}{covariates}
}
\value{
expit under the null model
}
\description{
Expit function under the null model
}
