% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sl_theta}
\alias{sl_theta}
\title{Initial parameter estimates under the null}
\usage{
sl_theta(Y, S, X, id.t, weights = NULL)
}
\arguments{
\item{Y}{Binary response variable}

\item{S}{Continuous surrogate}

\item{X}{Covaraites with intercept column design matrix}

\item{id.t}{Row id of labeled data.}

\item{weights}{Weights of the individual in the sample, default treat data equally.}
}
\value{
Initial estimates of the parameters
}
\description{
Initial parameter estimates under the null
}
