% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ssl_theta}
\alias{ssl_theta}
\title{Final parameter estimates under the null}
\usage{
ssl_theta(
  Y,
  X,
  S,
  Z,
  id.t,
  weights = NULL,
  full_eval = TRUE,
  NULL_nlog_like,
  nit,
  distri
)
}
\arguments{
\item{Y}{Binary response variable}

\item{X}{Covariates}

\item{S}{Continuous surrogate}

\item{Z}{Covaraites with intercept column design matrix}

\item{id.t}{Row id of labeled data.}

\item{weights}{Weights of the individual in the sample, default treat data equally.}

\item{full_eval}{Full optimization iteration, default to be TRUE.}

\item{NULL_nlog_like}{Function to be optimized, predefined as negative log-likelihood function under the null}

\item{nit}{Number of iteration for optimization if full_NR_evaluation is FALSE.}

\item{distri}{Distribution of S|Y, either "normal" or "beta"}
}
\value{
Final estimates of the parameters
}
\description{
Final parameter estimates under the null
}
